Added TilesModule.
A procedural module that generates real tiles. It does so by subdividing
space into 3d "boxes": If a points coordinate is close enough to the
next integer, the module returns 1 and 0 otherwise. This, in fact, tests
whether a point is close enough to its next "integer plane".

Note that there's a separate "2d mode": If enabled, it omits the test
for the Z-axis. That's necessary (or at least nice to have) if the
module is used in a 2d procedural texture. For now, the 2d mode has to
be enabled explicitely.