doKillLight(): Don't test for a specific name.
Test for the type of the object. This is OK since there's only one
single directional light in the initical scene.

This change ensures compatibility with a patch that may be applied
upstream. That patch, in turn, adds " (Point Light)" or " (Directional
Light)" to new lights -- but those strings are translated according to
the locale. As there's only one light anyway, we can simply check for
the object type.